# Optimize Docker page performance through algorithmic improvements