@Squidly271 2025-06-12
Had a look at this, and the caching is loading every single template from /templates-user.  

If the user has lots of templates they previously had installed, this results in many unnecessary reads.  Why not just cache the list of templates and then in getTemplateValue another cache is used for the template itself.  That way, only a template that has the value requested would get loaded on first request and cached for all others and everything not installed wouldn't wind up being cached

Should wind up being the same improvement if not better