@Squidly271 2025-06-13
I think you're over thinking this.

getRepositoryMap reads every .xml and even worse it's not caching it for subsequent reads.  With luck, the first cached file will have all the information required (eg: my-acrobat.xml is only thing installed and there's ton's of other xml's present)


All I'm suggesting is that in getTemplateValue you cache things as it goes.

EG: (not complete)
`public function getTemplateValue($Repository, $field, $scope='all',$name='') {
    foreach ($this->getTemplates($scope) as $file) {
        if (  ! isset($cachedFile[$Repository]) ) {
          $doc = new DOMDocument();
          $doc->load($file['path']);
          $cachedFile[$Repository] = $doc;
        } else $doc = $cachedFile[$Repository];

        if ($name) {
            if ($doc->getElementsByTagName('Name')->item(0)->nodeValue !== $name) continue;
        }
        $TemplateRepository = DockerUtil::ensureImageTag($doc->getElementsByTagName('Repository')->item(0)->nodeValue??'');
        if ($TemplateRepository && $TemplateRepository==$Repository) {
            $TemplateField = $doc->getElementsByTagName($field)->item(0)->nodeValue??'';
            return trim($TemplateField);
        }
    }
    return null;
}`

Trying to cache everything in advance means you're going to read more than is potentially required, and instead IMO cache as you read things.  Best case scenario its tons faster.