@Joly0 2025-06-14
Hey @Squidly271 thanks for the feedback. Yes, i was probably over thinking this heavily and got also side tracked by what rabbitai suggested. I follow that rabbit hole (lol) and over complicated this. I simplified this now a bit again and hopefully implemented the caching as you meant. If not, then i guess someone else should take a look at this. I tried a few different approaches to the general problem of this whole problem and havent found any other good solution that wouldnt worsen the performance. Atleast now the performance is not worse than with my initial first commit.