@Joly0 2025-06-26
> Hi! I wanted to let you know that we will be soon be rebuilding the Docker listing page from scratch, so it probably doesn't make sense to focus on optimizing the current version right now. We do appreciate your efforts though!

Unfortunate, would have been awesome if I could have contributed to unraid with a PR that is actually noticeable, but totally understandable that you decide not to merge it. Thanks for letting me know :D